rename func-nodes
